refactor: JPA 레포지토리 인터페이스와 구현체 통합
- 단순 위임만 하는 구현체를 만들어 줄 필요가 없어 복잡성이 줄어든다.

- Repository와 JpaRepository 사이에 로직이 존재하는 경우 해당 구조는
사용하기 어렵다.
