+++ src/main/java/roomescape/member/repository/MemberJpaRepository.java
@@ -1,10 +1,9 @@
-package roomescape.member.repository.jpa;
+package roomescape.member.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import roomescape.member.domain.Member;
 
-public interface MemberJpaRepository extends JpaRepository<Member, Long> {
-
+public interface  MemberJpaRepository extends JpaRepository<Member, Long>, MemberRepository {
     Optional<Member> findByEmailAndPassword(String email, String password);
 }
