+++ src/main/java/roomescape/reservation/repository/ReservationTimeJpaRepository.java
@@ -1,10 +1,9 @@
-package roomescape.reservation.repository.jpa;
+package roomescape.reservation.repository;
 
 import java.time.LocalTime;
 import roomescape.reservation.domain.ReservationTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationTimeJpaRepository extends JpaRepository<ReservationTime, Long> {
-
+public interface ReservationTimeJpaRepository extends JpaRepository<ReservationTime, Long>, ReservationTimeRepository {
     boolean existsByStartAt(LocalTime reservationTime);
 }
