+++ src/main/java/roomescape/reservation/repository/ThemeJpaRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository.jpa;
+package roomescape.reservation.repository;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -8,7 +8,7 @@ import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.ThemeName;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThemeJpaRepository extends JpaRepository<Theme, Long> {
+public interface ThemeJpaRepository extends JpaRepository<Theme, Long>, ThemeRepository {
 
     boolean existsByName(ThemeName name);
