--- src/main/java/roomescape/reservation/repository/jpa/ReservationTimeRepositoryImpl.java
@@ -1,43 +1,0 @@
-package roomescape.reservation.repository.jpa;
-
-import java.time.LocalTime;
-import java.util.List;
-import java.util.Optional;
-import org.springframework.stereotype.Repository;
-import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.repository.ReservationTimeRepository;
-
-@Repository
-public class ReservationTimeRepositoryImpl implements ReservationTimeRepository {
-
-    private final ReservationTimeJpaRepository reservationTimeJpaRepository;
-
-    public ReservationTimeRepositoryImpl(ReservationTimeJpaRepository reservationTimeJpaRepository) {
-        this.reservationTimeJpaRepository = reservationTimeJpaRepository;
-    }
-
-    @Override
-    public ReservationTime save(final ReservationTime reservationTime) {
-        return reservationTimeJpaRepository.save(reservationTime);
-    }
-
-    @Override
-    public List<ReservationTime> findAll() {
-        return reservationTimeJpaRepository.findAll();
-    }
-
-    @Override
-    public Optional<ReservationTime> findById(final long id) {
-        return reservationTimeJpaRepository.findById(id);
-    }
-
-    @Override
-    public boolean existsByTime(final LocalTime reservationTime) {
-        return reservationTimeJpaRepository.existsByStartAt(reservationTime);
-    }
-
-    @Override
-    public void deleteById(long id) {
-        reservationTimeJpaRepository.deleteById(id);
-    }
-}
