+++ src/test/java/roomescape/integration/TimeServiceIntegrationTest.java
@@ -23,7 +23,6 @@ import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.jpa.ReservationRepositoryImpl;
-import roomescape.reservation.repository.jpa.ReservationTimeRepositoryImpl;
 import roomescape.reservation.service.ReservationTimeService;
 import roomescape.reservation.service.dto.AvailableTimeInfo;
 import roomescape.reservation.service.dto.ReservationTimeCreateCommand;
@@ -31,7 +30,7 @@ import roomescape.reservation.service.dto.ReservationTimeInfo;
 
 @ActiveProfiles("test")
 @DataJpaTest
-@Import(value = {ReservationTimeRepositoryImpl.class, ReservationRepositoryImpl.class})
+@Import(value = ReservationRepositoryImpl.class)
 @DirtiesContext(classMode = ClassMode.BEFORE_CLASS)
 @Transactional
 @Sql(value = {"/test-data.sql"}, executionPhase = ExecutionPhase.BEFORE_TEST_CLASS)
